Added more parameters to the model in approxsq.
Modified mutation in GA2 so it makes a 20% change in the gene pool, also
added a small chance for the multiplication value to be negative.

Modified gatest, increased the range of values on the starting
generation and the number of generations. The best fitness reached is: 24.631464266778497
That is very close to the global optimum.